requirements.txt
face_recognition
opencv-python
numpy
Pillow
scikit-image
moviepy
pyheif
pillow-heif
